MB-25123 determine if nodes are in time sync with orchestrator
We determine if times are in sync by having each node send a
time-stamped time offset request to the orchestrator. The
orchestrator sends back a reply consisting of the node's request
along with the orchestrator's timestamp.  When the node receives
the reply, it compares the request timestamp, the reply timestamp
and the time the reply was received to determine the reply/response
round trip time and the time delta between the node and the
orchestrator.  If that delta exceeds a threshold, an alert is
issued.

Change-Id: I2a39d7fbf0332509854b3bfb09254beaa0fae66e
Reviewed-on: http://review.couchbase.org/c/ns_server/+/130466
Tested-by: Sam Cramer <sam.cramer@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>