MB-42664 - Long running query stats are counters, need 1m rate
The counts of long running queries, coming from cbq-engine, are in the
form of a counter that only resets when the server is
restarted. Traditionally we have shown these values as a per-minute
rate. It doesn't make sense to show long-running queries, which can
take many seconds, as a per-second rate. This change updates those
stats so they are displayed the same as 6.6, as a per-minute rate.

Change-Id: I5b4f01b5f0130e3b86920452830a35ca2028968f
Reviewed-on: http://review.couchbase.org/c/ns_server/+/149663
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>