ACE wasn't loading properly if the browser was refreshed on a non-ace page.
Weirdly, ACE worked just fine if you reloaded the browser on an
Angular 8 page using it (Doc Editor, Query Workbench), but if you
reloaded on a different page, like Logs, and then navigated to
Documents or Query Workbench, the ACE language customizations didn't
work.

It turns out that ui-ace.js (which was AngularJS), included an import
which was missing from ngx-ace.js (the Angular 8 wrapper for
Ace). This change adds the import, and now all works properly.

Change-Id: Idccedd7d7a362e21f2bcbcfeae83afd8b652b384
Reviewed-on: http://review.couchbase.org/c/ns_server/+/163234
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>