MB-50509: Auto-Compaction form: allow disabling size fields
When a database/view fragmentation size field is disabled by unchecking
the checkbox next to it, UI needs to send 0 to
"controller/setAutoCompaction" endpoint to reset the value.
Server endpoint "settings/autoCompaction" returns
size: "undefined" and the UI disables the checkbox in this case.

Change-Id: Iddf901be3068c87fb605bd04401bcad751e230b3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169750
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>