MB-50299 Move memory quota dialog to mn_admin_config
Since the setup wizard has been upgraded, we weren't
broadcasting the event to open the new service
dialog.

The logic for listening to the event has been
moved to mn_admin_config, so that the '$scope.on'
is able to listen to the broadcasted event
from the new 'join.cluster' component.

Change-Id: Ie7a1f35a8e04ad5918f474f8b3e0677717eb54a8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169922
Tested-by: Matthew Dawber <matthew.dawber@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>