MB-40229 - allow deleting either single index replica or all replicas
In prior releases, there was no way in the UI to drop a single replica
of a replicated index. Instead, the "Drop Index" button on the Indexes
page would drop every replica, even if it was not obvious to the user
that the index was replicated.

The previous fix for this issue caused only the replica to be dropped
when the user clicked "Drop Index" for a replicated index. It was
requested that the user have the option to either drop the replica, or
drop the entire index.

This change provides two buttons for replicated indexes, one labelled
"Drop This Replica", and the other labelled "Drop All Replicas".

Change-Id: Ieecf21d86f9d889ae5571fe41eb2737ec0449088
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170967
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>