MB-61229 - fix TTL tooltip which had information reversed.
In CB7.6, a collection TTL of 0 indicates that the value for TTL
should be inherited from the parent bucket (if any).

A collection TTL of -1 indicates that TTL should be disabled on the
collection, regardless of any TTL on the parent bucket.

The tooltips for TTL when creating or editing collections had these
reversed.

Change-Id: I02dffc4b061b2fa7e0f1f6c80bab982d94cb20fd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210039
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Restriction Checker