MB-60780 - Disallow un-checking "share usage" in CE
When creating a cluster, the checkbox for "Share usage information
with Couchbase" is disabled in CE, so people are required to share
information in exchange for getting CE for free. However, this can
also be changed in the Settings page. This change disables the
checkbox in CE, preventing the user from un-checking it.

Change-Id: I1a64d5f2ebe2bef015ef1cad40e19808b47abcde
Reviewed-on: https://review.couchbase.org/c/ns_server/+/211083
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>