MB-62693 - UI to allow changing services running on nodes of a cluster
With Morpheus we get a new API that permits us to change the service
topology of a cluster and rebalance in one step. (However, the kv
service cannot be added or removed from existing nodes.) This change
adds a UI to permit users to modify the topology.

There are several special cases that must be handled. The option is
not available on mixed clusters, or on CE clusters, or, when any node
is unhealthy.

The memory_quota_dialog is used when adding new nodes to the cluster,
if a new service is added. This dialog is also used whenever the user
uses this function to add a new service to existing nodes. There were
some minor tweaks to the language in this dialog, as well as adding
support for the Query service (which does have quota setting). In the case of changing services on existing nodes, I changed the quota dialog to show the quotas for all running services, not just the new ones, because assigning memory to a new service may be impacted by allocations to existing services.

Change-Id: I6fda23a842af9e02b19e1628eaa9d892fde21901
Reviewed-on: https://review.couchbase.org/c/ns_server/+/216314
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Eben Haber <eben@couchbase.com>