MB-62919 - fix UI checkbox issue with File Transfer Based Rebalance
Under Advanced Index Settings, the checkbox for "Enable File Transfer
Based Rebalance" was deliberately created as a one-way option, the
user could enable it but never disable it with the UI.

It was decided that this behavior isn't ideal. Instead, the user will
be permitted to disable the option, but they will be shown a warning
message if they try to do so.

Change-Id: I40808d5cb46c98e90808b124fc9d7cbe8b89c385
Reviewed-on: https://review.couchbase.org/c/ns_server/+/216859
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>