MB-64080 - Make config dialog password placeholders more accurate.
When using the UI wizard to configure a new cluster, or connect a new
node to an existing cluster, the password field has a placeholder
saying, "6-characters minimum". However, that might not be the case if
the cluster is configured with a different password policy.

This change updates the placeholders as suggested in the ticket to be
more accurate.

Change-Id: I593bdf43f288e9d4a2d6f76cb9ccc4cd93e7ccd3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219013
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Eben Haber <eben@couchbase.com>