MB-642319 - add checkbox to enable DEK bucket rotation, lifetime
The goal of this bug is to make the DEK controls in the buckets page
similar to the DEK controls on the encryption page, with a checkbox to
enable/disable rotation or lifetime specification, and a text field to
specify the values. Before now, the buckets panel only had a text
field, and the user had to specify 0 to disable the setting.

Change-Id: Ib4ea9667f93ec0762558542eb1bfd36ff78c8ecf
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219703
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>