MB-63209 - UI to toggle disk storage / congfigure disk space guardrails
Morpheus includes guard rails to prevent data writes when a disk is
full more than a set amount. If a disk gets too full, it can become
impossible to rebalance or recover a node.

The UI is simple: a check box to enable/disable the feature, and a
text field to allow the set full value to be changed from its default
of 85%.

Feature was validated through UI testing running locally.

Change-Id: I46a62a1b2c4bcae245f5a56632fd726b4e392ddd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220892
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>