MB-53907 - allow user to see long bucket names in index view
In the list of indexes, long values may be truncated. The index name
already had a tag to display the full name on hover, but the same was
not true of they keyspace name. This change adds a title tag so the
keyspace name appears when the user hovers their mouse over it.

Change-Id: Ica2dc8d9720f41fb0d69c6c51a5fbe2e7daaea52
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220967
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>