MB-64608: KMIP Secret Type.
Adds a new secret type - KMIP.
The new type of secret can be chosen in the
Add Secret dialog with its specific fields.
The list of keys are displayed when a KMIP
secret is expanded from the secrets list.
Rotate button is hidden fot KMIP secrets
because rotation is not supported for them.

Change-Id: I39fa1388ffe55e44f302c5fcdc158f903bc4ec5c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221292
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>